Update feed (#41)
* update get children by parent id

* update feed

* update feed

* update feed

* update feed

* update feed

* update feed

---------

+++ src/main/java/app/mobile/BK_Kindergarten/children/Children.java
@@ -15,7 +15,7 @@ public class Children {
     @Id
     @Column(name = "child_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name = "name")
     private  String name;
@@ -36,5 +36,5 @@ public class Children {
     private String character;
 
     @JoinColumn(name = "parent_id")
-    private Long  parent_id;
+    private int  parent_id;
 }
